+++ servant-auth-server/test/Servant/Auth/ServerSpec.hs
@@ -68,7 +68,6 @@ spec = do
   authSpec
   cookieAuthSpec
   jwtAuthSpec
-  throwAllSpec
   basicAuthSpec
 
 ------------------------------------------------------------------------------
@@ -381,24 +380,6 @@ basicAuthSpec = describe "The BasicAuth combinator"
 
 -- }}}
 ------------------------------------------------------------------------------
--- * ThrowAll {{{
-
-throwAllSpec :: Spec
-throwAllSpec = describe "throwAll" $ do
-
-  it "works for plain values" $ do
-    let t :: Either ServerError Int :<|> Either ServerError Bool :<|> Either ServerError String
-        t = throwAll err401
-    t `shouldBe` throwError err401 :<|> throwError err401 :<|> throwError err401
-
-  it "works for function types" $ property $ \i -> do
-    let t :: Int -> (Either ServerError Bool :<|> Either ServerError String)
-        t = throwAll err401
-        expected _ = throwError err401 :<|> throwError err401
-    t i `shouldBe` expected i
-
--- }}}
-------------------------------------------------------------------------------
 -- * API and Server {{{
 
 type API auths
@@ -460,29 +441,30 @@ instance FromBasicAuthData User where
 -- have to add it
 type instance BasicAuthCfg = JWK
 
-appWithCookie :: AreAuths auths '[CookieSettings, JWTSettings, JWK] User
+appWithCookie :: AreAuths auths '[CookieSettings, AuthErrorHandler, JWTSettings, JWK] User
   => Proxy (API auths) -> CookieSettings -> Application
 appWithCookie api ccfg = serveWithContext api ctx $ server ccfg
   where
-    ctx = ccfg :. jwtCfg :. theKey :. EmptyContext
+    ctx = ccfg
+       :. authErrorHandler401
+       :. jwtCfg
+       :. theKey
+       :. EmptyContext
 
 -- | Takes a proxy parameter indicating which authentication systems to enable.
-app :: AreAuths auths '[CookieSettings, JWTSettings, JWK] User
+app :: AreAuths auths '[CookieSettings, AuthErrorHandler, JWTSettings, JWK] User
   => Proxy (API auths) -> Application
 app api = appWithCookie api cookieCfg
 
 server :: CookieSettings -> Server (API auths)
 server ccfg =
-    (\authResult -> case authResult of
-        Authenticated usr -> getInt usr
-                        :<|> postInt usr
-                        :<|> getHeaderInt
+    (\user -> getInt user
+         :<|> postInt user
+         :<|> getHeaderInt
 #if MIN_VERSION_servant_server(0,15,0)
-                        :<|> return (S.source ["bytestring"])
+         :<|> return (S.source ["bytestring"])
 #endif
-                        :<|> raw
-        Indefinite -> throwAll err401
-        _ -> throwAll err403
+         :<|> raw
     )
     :<|> getLogin
     :<|> getLogout